Fixed return type from NextDouble.

+++ Extensions.cs
@@ -21,7 +21,7 @@ namespace XnaGameLib
 			return min + (float) (random.NextDouble() * (max - min));
 		}
 
-		public static float NextDouble(this Random random, double min, double max)
+		public static double NextDouble(this Random random, double min, double max)
 		{
 			return min + random.NextDouble() * (max - min);
 		}
